Solved a few latent issues with lock double-init/double-free. LibodbTest's minimal case compiles on Windows now.
The locking issues weren't noticed until trying to use CPP11 threads since the pthreads locking constructs didn't need new/free operations.

+++ src/include/utility.hpp
@@ -53,12 +53,12 @@ namespace libodb
 
     inline uint32_t len_v(void* rawdata)
     {
-        return strlen((const char*)rawdata);
+        return (uint32_t)strlen((const char*)rawdata);
     }
 
     inline bool search(std::vector<void*>* marked, void* addr)
     {
-        int32_t start = 0, end = marked->size() - 1, midpoint = (start + end) / 2;;
+        size_t start = 0, end = (marked->size() - 1), midpoint = (start + end) / 2;;
         int64_t c;
 
         while (start <= end)
